tests/wpt/web-platform-tests/html/semantics/forms/the-output-element/output-setcustomvalidity.html

<!DOCTYPE HTML>
<title>output setCustomValidity</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<output id='output_test'></output>

<script>

test(() => {
  let elem = document.getElementById("output_test");
  assert_false(elem.validity.customError);
  elem.setCustomValidity("custom error");
  assert_true(elem.validity.customError);
}, "output setCustomValidity is correct")

</script>
